Demuxer: Fix clang/-Werror and 32-bit portability issues
- ISODemuxer/wav: comparisons of a size_t/uint32_t against 2^32 and
  INT64_MAX are tautological where size_t is 32-bit (i386); cast to
  uint64_t so the bounds checks stay well-formed (and meaningful on
  32-bit off_t platforms) instead of -Werror,-Wtautological-compare.
- DemuxedStream: drop std::move on a function return value
  (-Werror,-Wpessimizing-move).
- FLACDemuxer: remove unused m_frame_index_complete field
  (-Werror,-Wunused-private-field).

+++ src/demuxer/iso/ISODemuxer.cpp
@@ -55,8 +55,10 @@ void ISODemuxer::InitializeMemoryManagement() {
     // Enable lazy loading based on system memory
     MemoryTracker::MemoryStats stats = memoryTracker.getStats();
     if (stats.total_physical_memory > 0) {
-        // Enable lazy loading if system has less than 4GB RAM
-        bool enableLazyLoading = (stats.total_physical_memory < 4ULL * 1024 * 1024 * 1024);
+        // Enable lazy loading if system has less than 4GB RAM. Cast to 64-bit so
+        // the comparison is well-formed where size_t is 32-bit (e.g. i386), where
+        // it is simply always true rather than a tautological-compare error.
+        bool enableLazyLoading = (static_cast<uint64_t>(stats.total_physical_memory) < 4ULL * 1024 * 1024 * 1024);
         if (sampleTables) {
             sampleTables->EnableLazyLoading(enableLazyLoading);
         }
